backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["cosmetic master <djangoreactdev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
langchain = "^0.0.325"
openai = "^0.28.1"
python-dotenv = "^1.0.0"
tiktoken = "^0.5.1"
faiss-cpu = "^1.7.4"
numexpr = "^2.8.7"
jq = "^1.6.0"
wget = "^3.2"

[tool.poetry.dependencies.qdrant-client]
version = ">=1.6.4,<2.0.0"
python = ">=3.10,<3.13"

[tool.poetry.dependencies.numpy]
version = "^1.26.1"
python = ">=3.10,<3.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
